--- frontend/src/features/openAi/prompts/publicationCheckPrompt.md
@@ -1,210 +1,0 @@
-# PR TIMES プレスリリース掲載基準 - AI向け構造化ガイド
-
-## 🎯 基本コンセプト
-
-### プラットフォームの性質
-
-- **目的**: 報道関係者が「ニュース」として活用できる情報の流通
-- **判断基準**: 新規情報性 × ニュースバリュー
-- **審査体制**: システム自動判定 + CR チーム全件目視審査
-
-### 掲載可能な契機
-
-- ✅ 新サービス開始、限定商品発売、キャンペーン、イベント、業務提携
-- ✅ 不祥事や事故などのネガティブ事象も含む（透明性重視）
-
----
-
-## ✅ 必須要件（MUST）
-
-### 1. 新規公開情報の包含
-
-- **必須**: 必ず新規公開情報を含むこと
-- **NG例**: 既存製品の再告知、実施中キャンペーンの複数回発信
-
-### 2. 発信主体と登録企業の関係性
-
-- **必須**: 関係性が明白であること
-- **NG例**: 所在不明な団体からの発信
-
----
-
-## ❌ 配信不可コンテンツ
-
-### 法令・公序良俗違反
-
-- 薬機法・景品表示法抵触の誇大広告
-- 違法薬物使用助長
-- 自殺・自傷行為関連
-- 詐欺商法
-
-### 誹謗中傷・差別表現
-
-- 民族、宗教、人種、性別、年齢に関する差別
-- 悪意のある発信
-- フェイクニュース拡散
-- 一方的な他者批判
-
----
-
-## 📋 特定ジャンル掲載基準
-
-### 商品の再販・再入荷
-
-- **原則**: NG（日常業務活動のため）
-- **例外**: 期間限定品、異例の完売商品、販売方針変更を伴う場合
-
-### メディア掲載実績
-
-- **判定**: 完全NG
-- **理由**: 報道主体が「メディア側」にあるため
-
-### インタビュー・対談記事
-
-- **原則**: NG（企業活動の開始・成果ではないため）
-- **例外**: プロモーション企画の一環で、その旨が本文で確認できる場合
-
-### 調査レポート
-
-- **条件**: 以下5項目の明記が必須
-  1. 調査期間
-  2. 調査機関（調査主体）
-  3. 調査対象
-  4. 有効回答数（サンプル数）
-  5. 調査方法（集計・算出方法）
-
-### 婚活イベント
-
-- **条件**: 以下2点を同時満足
-  1. 身分証明書提示または本人確認が必須
-  2. その旨がプレスリリース本文またはイベント公式サイトに記載
-
-### 漫画・アニメ・ゲーム関連
-
-- **NG要素**:
-  - 18歳未満向けコンテンツでの性的表現
-  - 性器、行為、裸に関する具体的名称・画像
-  - 露骨に示唆する表現
-
----
-
-## 🚫 掲載不可業種
-
-### 医療・美容系
-
-- 医療行為を伴う内容
-- 美容・エステ・整体・気功・カイロプラクティック・はり・灸での医療行為表現
-- **関連法令**: 医療広告ガイドライン、薬機法
-
-### その他
-
-- 全日本遊技事業協同組合連合会注意事項該当内容
-- 反社会的勢力関連
-- 公序良俗違反イベント主催・運営
-
----
-
-## 🖼️ 画像・ビジュアル基準
-
-### ビフォー・アフター画像
-
-- **判定**: 原則NG（薬機法抵触リスク）
-- **例外**: 客観的事実を示す研究結果発表
-
-### 性的・猥褻表現画像
-
-- **NG基準**:
-  - アダルトコンテンツを想起させる画像
-  - 肌の露出が多い画像
-  - 胸が「半カップ以上隠れていない」画像
-  - 着衣でもアダルトを暗示する画像
-
----
-
-## 🏢 企業登録審査
-
-### 必須条件
-
-1. **反社チェック**: 反社会的勢力ではないこと
-2. **事業確認**: 会社住所、代表者名、事業内容が確認可能
-3. **基準適合**: 事業内容がコンテンツ基準を満たすこと
-
-### 審査フロー
-
-- **所要時間**: 最長1営業日以内
-- **優先対応**: サポート＆サクセスデスクへの相談で可能
-- **重複確認**: 同一企業の既存アカウント確認
-
-### アカウント管理
-
-- **原則**: 正式な法人名・団体名で登録
-- **例外**: 事業部・ブランド別の複数アカウント可能
-
----
-
-## 📝 掲載後管理
-
-### 訂正・取り下げ
-
-- **企業側**: 自己責任で内容訂正可能
-- **制限**: パートナーメディア転載分の訂正は保証外
-- **取り下げ**: PR TIMESへの依頼が必要
-
-### PR TIMES権限
-
-- **強制訂正・取り下げ**: 規約・基準違反時に予告なく実行
-- **損害責任**: PR TIMESは一切負わない
-- **誤字脱字**: 不利益判断時に予告なく訂正
-
----
-
-## ⚖️ 規約構造
-
-### 階層構造
-
-- **基本規約**: 全利用者共通の大原則
-- **個別規約**: 特定利用者・形態向け詳細ルール
-- **優先順位**: 矛盾時は個別規約が優先
-
-### 関連法令
-
-- 薬機法（旧薬事法）
-- 景品表示法
-- 医療広告ガイドライン
-- 全日本遊技事業協同組合連合会注意事項
-
----
-
-## 📊 クイックリファレンス表
-
-| カテゴリ       | 項目               | 判定   | 条件・備考             |
-| -------------- | ------------------ | ------ | ---------------------- |
-| **基本要件**   | 新規情報           | MUST   | 再告知・反復配信NG     |
-| **基本要件**   | 発信主体関係性     | MUST   | 登録企業との関係明確化 |
-| **基本要件**   | 誤字脱字           | 条件付 | AIの判定ではMust       |
-| **コンテンツ** | 誹謗中傷・差別     | NG     | 完全禁止               |
-| **コンテンツ** | 法令違反           | NG     | 薬機法・景表法等       |
-| **ジャンル**   | メディア掲載実績   | NG     | 完全禁止               |
-| **ジャンル**   | 商品再販           | 条件付 | 例外的に許可の場合あり |
-| **ジャンル**   | 調査レポート       | 条件付 | 5項目明記必須          |
-| **ジャンル**   | 婚活イベント       | 条件付 | 本人確認必須明記       |
-| **画像**       | ビフォー・アフター | NG     | 研究発表は例外         |
-| **画像**       | 性的表現           | NG     | 完全禁止               |
-| **業種**       | 医療・美容         | NG     | 医療行為表現禁止       |
-
----
-
-## 💡 AI処理時の注意点
-
-### 判断の優先順位
-
-1. **法令遵守**: 薬機法、景品表示法、医療広告ガイドライン
-2. **新規性**: 過去配信内容との重複チェック
-3. **ニュース価値**: 報道機関が扱う価値があるか
-4. **社会性**: 公序良俗、差別表現の排除
-
-### 境界ケースの処理
-
-- **グレーゾーン**: PR TIMES審査チームの裁量権を考慮
-- **例外処理**: 文脈と社会情勢を総合判断
-- **更新性**: 基準は社会動向により変動する可能性
