--- frontend/src/features/openAi/services/client.ts
@@ -1,10 +1,0 @@
-"server-only";
-
-import OpenAI from "openai";
-
-/**
- * @description OpenAIのclient
- */
-export const openAiClient = new OpenAI({
-  apiKey: process.env.OPENAI_API_KEY!,
-});
