--- frontend/src/features/openAi/types/FileNameEnum.ts
@@ -1,6 +1,0 @@
-/**
- * @description prompt用のMarkdown Fileの名前をEnumで管理
- */
-export enum fileNameEnum {
-  publicationCheck = "publicationCheckPrompt.md",
-}
